refactor: authLink, interestsTag 파일명 변경 (#60)
