+++ app/(auth)/_component/interests-tag.tsx
@@ -1,0 +1,18 @@
+import Button from "@/components/Button";
+import TagGroup from "@/components/TagGroup/TagGroup";
+import { languageOptions, levelOptions, problemOptions } from "@/config/study";
+
+const InterestsTag = () => {
+  return (
+    <div className="flex flex-col gap-[78px] max-w-screen-md items-center">
+      <div className="flex flex-col gap-[60px] px-4">
+        <TagGroup label="언어" options={languageOptions} />
+        <TagGroup label="풀이 사이트" options={problemOptions} />
+        <TagGroup label="레벨" options={levelOptions} />
+      </div>
+      <Button className="w-[55%] m-auto items-center py-3">확인</Button>
+    </div>
+  );
+};
+
+export default InterestsTag;
